refactor(schema): lift fixture loaders into shared testHelpers
Extract toCamel/camelizeKeys/loadFixture + FIXTURE_DIR out of
lessonContent.contract.test.ts into src/schema/testHelpers.ts so the
Phase-2 lesson-content renderer tests can re-use the exact same
snake->camel + zod-parse path (single implementation, lock-step with
the frozen contract). Contract test unchanged behaviourally (38 pass).
